progress on making the ios SDK docs complete
