--- StitchCore-iOS/DocsSources/Abstracts/Overview.md
@@ -1,1 +1,0 @@
-This will probably be the README.
