--- StitchCore-iOS/DocsSources/Pages/Getting Started.md
@@ -1,3 +1,0 @@
-# Getting Started
-
-This will be a page about how to get started with the SDK.
