+++ StitchCore-iOS/README.md
@@ -1,0 +1,11 @@
+# [![](https://www.rosehosting.com/blog/wp-content/uploads/2014/12/mongodb-50x50.png) MongoDB Stitch](https://mongodb.com/cloud/stitch) iOS SDK
+
+[![Join the chat at https://gitter.im/mongodb/stitch](https://badges.gitter.im/mongodb/stitch.svg)](https://gitter.im/mongodb/stitch?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge&utm_content=badge) ![iOS](https://img.shields.io/badge/platform-iOS-blue.svg) [![Swift 4.0](https://img.shields.io/badge/swift-4.0-orange.svg)](https://developer.apple.com/swift/) ![Apache 2.0 License](https://img.shields.io/badge/license-Apache%202-lightgrey.svg) [![Cocoapods compatible](https://img.shields.io/badge/pod-v1.0.0-ff69b4.svg)](#Cocoapods)
+
+[MongoDB Stitch Users - Google Group](https://groups.google.com/d/forum/mongodb-stitch-users)
+
+[MongoDB Stitch Announcements - Google Group](https://groups.google.com/d/forum/mongodb-stitch-announce)
+
+StitchCore_iOS is the primary Swift module for the MongoDB Stitch iOS SDK. TODO: Update this to reflect usage of the new SDK.
+
+See the [Getting Started](/Getting%20Started.html) page for instructions on how to begin integrating MongoDB Stitch with your iOS apps.
