+++ StitchCore/Sources/StitchCore/Auth/CoreAuthProviderClient.swift
@@ -1,9 +1,30 @@
-/** :nodoc: */
+/**
+ * :nodoc:
+ * The class from which all Core auth provider clients inherit.
+ */
 open class CoreAuthProviderClient {
+    // MARK: Properties
+
+    /**
+     * The name of the authentication provider.
+     */
     public let providerName: String
+
+    /**
+     * The `StitchRequestClient` used by the client to make requests.
+     */
     public let requestClient: StitchRequestClient
+
+    /**
+     * The `StitchAuthRoutes` object representing the authentication API routes on the Stitch server.
+     */
     public let authRoutes: StitchAuthRoutes
 
+    // MARK: Initializer
+
+    /**
+     * A basic initializer, which sets the provider client's properties to the values provided in the parameters.
+     */
     init(withProviderName providerName: String,
          withRequestClient requestClient: StitchRequestClient,
          withAuthRoutes authRoutes: StitchAuthRoutes) {
