+++ StitchCore/Sources/StitchCore/Auth/Internal/CoreStitchAuth.swift
@@ -1,25 +1,52 @@
 import Foundation
 import ExtendedJSON
 
-private enum AuthKey: String {
-    case options
-    case device
-}
-
 /**
- * synchronization in this class happens around the authInfo and currentUser objects such that
- * access to them is 1. always atomic and 2. queued to prevent excess token refreshes.
+ * The core class that holds and manages Stitch user authentication state. This class is meant to be inherited.
  *
- * @param <TStitchUser>
+ * - important: synchronization in this class happens around the authInfo and currentUser objects such that
+ *              access to them is 1. always atomic and 2. queued to prevent excess token refreshes.
+ *
+ * - typeparameters
+ *     - TStitchUser: The underlying user type for this `CoreStitchAuth`, which must conform to `CoreStitchUser`.
  */
 open class CoreStitchAuth<TStitchUser> where TStitchUser: CoreStitchUser {
+    // MARK: Stored Properties
+
+    /**
+     * The underlying authentication state of this `CoreStitchAuth`
+     */
     internal var authStateHolder = AuthStateHolder()
-    private var storage: Storage
+
+    /**
+     * The `Storage` object indicating where authentication information should be persisted.
+     */
+    internal var storage: Storage
+
+    /**
+     * The thread that will proactively refresh the access token at fixed intervals.
+     */
     private var refresherThread: Thread?
+
+    /**
+     * A `TStitchUser` object that represents the currently authenticated user, or `nil` if no one is authenticated.
+     */
     private var currentUser: TStitchUser?
 
+    /**
+     * The `StitchRequestClient` used by the `CoreStitchAuth` to make requests to the Stitch server.
+     */
     public let requestClient: StitchRequestClient
+
+    /**
+     * The `StitchAuthRoutes` object representing the authentication API routes of the Stitch server for the current
+     * app.
+     */
     public let authRoutes: StitchAuthRoutes
+
+    /**
+     * The getter and setter for authentication state, as represented by an `AuthInfo` object.
+     */
     public internal(set) var authInfo: AuthInfo? {
         get {
             return authStateHolder.authInfo
@@ -29,6 +56,13 @@ open class CoreStitchAuth<TStitchUser> where TStitchUser: CoreStitchUser {
         }
     }
 
+    // MARK: Initialization
+
+    /**
+     * Initializes the `CoreStitchAuth` with a request client, authentication API routes, and a `Storage` indicating
+     * where the `CoreStitchAuth` should persist authentication information. This initializer will start a `Thread`
+     * that will proactively refresh the access token at fixed intervals.
+     */
     public init(requestClient: StitchRequestClient,
                 authRoutes: StitchAuthRoutes,
                 storage: Storage) throws {
@@ -55,26 +89,50 @@ open class CoreStitchAuth<TStitchUser> where TStitchUser: CoreStitchUser {
         self.refresherThread?.start()
     }
 
+    /**
+     * Instantiates an access token refresher and begins its infinite loop.
+     *
+     * - important: Should only be called on a standalone non-main thread.
+     */
     @objc private func doRunAccessTokenRefresher() {
         AccessTokenRefresher<TStitchUser>(authRef: self).run()
     }
 
+    /**
+     * Cancels the access token refresher thread.
+     */
     deinit {
         refresherThread?.cancel()
     }
 
+    // MARK: Unimplemented Methods and Properties
+
+    /**
+     * Should return an `AnyStitchUserFactory` object, capable of constructing users of the `TStitchUser` type.
+     */
     open var userFactory: AnyStitchUserFactory<TStitchUser> {
         fatalError("not implemented")
     }
 
+    /**
+     * A method that will be called whenever an authentication event (logging in, logging out, linking) occurs.
+     */
     open func onAuthEvent() {
         fatalError("not implemented")
     }
 
+    /**
+     * A method that should return a BSON Document containing information about the current device.
+     */
     open var deviceInfo: Document {
         fatalError("deviceInfo must be implemented")
     }
 
+    // MARK: Computed Properties
+
+    /**
+     * Whether or not a user is currently logged in.
+     */
     public var isLoggedIn: Bool {
         // swiftlint:disable force_try
         return try! sync(self) {
@@ -83,6 +141,9 @@ open class CoreStitchAuth<TStitchUser> where TStitchUser: CoreStitchUser {
         }
     }
 
+    /**
+     * The currently authenticated user as a `TStitchUser`, or `nil` if no user is currently authenticated.
+     */
     public var user: TStitchUser? {
         // swiftlint:disable force_try
         return try! sync(self) {
@@ -91,6 +152,29 @@ open class CoreStitchAuth<TStitchUser> where TStitchUser: CoreStitchUser {
         }
     }
 
+    /**
+     * Returns whether or not the current authentication state has a meaningful device ID.
+     */
+    public var hasDeviceId: Bool {
+        return authInfo?.deviceId != nil
+            && authInfo?.deviceId != ""
+            && authInfo?.deviceId != "000000000000000000000000"
+    }
+
+    /**
+     * Returns the currently authenticated user's device ID, or `nil` is no user is currently authenticated, or if the
+     * device ID does not exist.
+     */
+    public var deviceId: String? {
+        return authInfo?.deviceId
+    }
+
+    // MARK: Authentication Actions
+
+    /**
+     * Authenticates the `CoreStitchAuth` using the provided `StitchCredential. Blocks the current thread until the
+     * request is completed.
+     */
     public func loginWithCredentialBlocking(withCredential credential: StitchCredential) throws -> TStitchUser {
         return try sync(self) {
             if !isLoggedIn {
@@ -108,6 +192,10 @@ open class CoreStitchAuth<TStitchUser> where TStitchUser: CoreStitchUser {
         }
     }
 
+    /**
+     * Links the currently logged in user with a new identity represented by the provided `StitchCredential. Blocks the
+     * current thread until the request is completed.
+     */
     public func linkUserWithCredentialBlocking(withUser user: TStitchUser,
                                                withCredential credential: StitchCredential) throws -> TStitchUser {
         return try sync(self) {
@@ -121,6 +209,12 @@ open class CoreStitchAuth<TStitchUser> where TStitchUser: CoreStitchUser {
         }
     }
 
+    /**
+     * Logs out the current user, and clears authentication state from this `CoreStitchAuth` as well as underlying
+     * storage. Blocks the current thread until the request is completed. If the logout request fails, this method will
+     * still attempt to clear local authentication state. This method will only throw if clearing authentication state
+     * fails.
+     */
     public func logoutBlocking() throws {
         guard isLoggedIn else { return }
 
@@ -135,57 +229,14 @@ open class CoreStitchAuth<TStitchUser> where TStitchUser: CoreStitchUser {
         try clearAuth()
     }
 
-    public var hasDeviceId: Bool {
-        return authInfo?.deviceId != nil
-            && authInfo?.deviceId != ""
-            && authInfo?.deviceId != "000000000000000000000000"
-    }
+    // MARK: Internal Methods
 
-    public var deviceId: String? {
-        return authInfo?.deviceId
-    }
-
-    // use this critical section to create a queue of pending outbound requests
-    // that should wait on the result of doing a token refresh or logout. This will
-    // prevent too many refreshes happening one after the other.
-    internal func tryRefreshAccessToken(reqStartedAt: TimeInterval) throws {
-        try sync(self) {
-            guard isLoggedIn, let accessToken = self.authStateHolder.accessToken else {
-                throw StitchError.requestError(withMessage: "logged out during request")
-            }
-
-            let jwt = try DecodedJWT.init(jwt: accessToken)
-            guard let issuedAt = jwt.issuedAt,
-                issuedAt.timeIntervalSince1970 < reqStartedAt else {
-                return
-            }
-            try refreshAccessToken()
-        }
-    }
-
-    /// NOTE: This method must be called within a lock
-    internal func refreshAccessToken() throws {
-        let response = try self.doAuthenticatedRequest(StitchAuthRequestBuilderImpl {
-            $0.useRefreshToken = true
-            $0.path = self.authRoutes.sessionRoute
-            $0.method = .post
-        }.build())
-
-        let newAccessToken = try JSONDecoder().decode(APIAccessToken.self,
-                                                   from: response.body!)
-
-        self.authInfo = self.authInfo?.refresh(withNewAccessToken: newAccessToken)
-
-        try self.authInfo?.write(toStorage: &self.storage)
-    }
-
-    private func attachAuthOptions(authBody: inout Document) {
-        authBody[AuthKey.options.rawValue] = [
-            AuthKey.device.rawValue: deviceInfo
-        ] as Document
-    }
-
-    // callers of doLogin should be synchronized before calling in.
+    /**
+     * Performs the logic of logging in this `CoreStitchAuth` as a new user with the provided credential. Can also
+     * perform a user link if the `asLinkRequest` parameter is true.
+     *
+     * - important: Callers of `doLogin` should be synchronized before calling in.
+     */
     private func doLogin(withCredential credential: StitchCredential, asLinkRequest: Bool) throws -> TStitchUser {
         let response = try self.doLoginRequest(withCredential: credential,
                                                asLinkRequest: asLinkRequest)
@@ -197,6 +248,29 @@ open class CoreStitchAuth<TStitchUser> where TStitchUser: CoreStitchUser {
         return user
     }
 
+    /**
+     * Enum representing the keys for additional auth options that may be attached to the body of the authentication
+     * request sent to the Stitch server on login or link.
+     */
+    private enum AuthKey: String {
+        case options
+        case device
+    }
+
+    /**
+     * Attaches authentication options to the BSON document passed in as the `authBody` parameter. Necessary for the
+     * the login request.
+     */
+    private func attachAuthOptions(authBody: inout Document) {
+        authBody[AuthKey.options.rawValue] = [
+            AuthKey.device.rawValue: deviceInfo
+            ] as Document
+    }
+
+    /**
+     * Performs the login request against the Stitch server. If `asLinkRequest` is true, a link request is performed
+     * instead.
+     */
     private func doLoginRequest(withCredential credential: StitchCredential,
                                 asLinkRequest: Bool) throws -> Response {
         let reqBuilder = StitchDocRequestBuilderImpl {
@@ -225,6 +299,10 @@ open class CoreStitchAuth<TStitchUser> where TStitchUser: CoreStitchUser {
         }.build())
     }
 
+    /**
+     * Processes the response of the login/link request, setting the authentication state if appropriate, and
+     * requesting the user profile in a separate request.
+     */
     private func processLoginResponse(withCredential credential: StitchCredential,
                                       forResponse response: Response) throws -> TStitchUser {
         guard let body = response.body else {
@@ -269,6 +347,9 @@ open class CoreStitchAuth<TStitchUser> where TStitchUser: CoreStitchUser {
         return self.currentUser!
     }
 
+    /**
+     * Performs a request against the Stitch server to get the currently authenticated user's profile.
+     */
     private func doGetUserProfile() throws -> StitchUserProfile {
         let response = try doAuthenticatedRequest(StitchAuthRequestBuilderImpl {
             $0.method = .get
@@ -283,6 +364,9 @@ open class CoreStitchAuth<TStitchUser> where TStitchUser: CoreStitchUser {
                                           data: decodedProfile.data)
     }
 
+    /**
+     * Performs a logout request against the Stitch server.
+     */
     @discardableResult
     private func doLogout() throws -> Response {
         return try self.doAuthenticatedRequest(StitchAuthRequestBuilderImpl {
@@ -292,6 +376,10 @@ open class CoreStitchAuth<TStitchUser> where TStitchUser: CoreStitchUser {
         }.build())
     }
 
+    /**
+     * Clears the `CoreStitchAuth`'s authentication state, as well as associated authentication state in underlying
+     * storage.
+     */
     internal func clearAuth() throws {
         try sync(self) {
             guard self.isLoggedIn else { return }
