+++ StitchCore/Sources/StitchCore/Auth/Internal/DeviceField.swift
@@ -1,10 +1,36 @@
 import Foundation
 
+/**
+ * An enum indicating the fields expected in a `CoreStitchAuth`'s `deviceInfo` BSON document.
+ */
 public enum DeviceField: String {
+    /**
+     * The current user's device ID.
+     */
     case deviceId
+
+    /**
+     * The name of the local application.
+     */
     case appId
+
+    /**
+     * The current version of the local application.
+     */
     case appVersion
+
+    /**
+     * The identifer for the current device's platform. (eg. "iOS")/
+     */
     case platform
+
+    /**
+     * The version of the current device's platform.
+     */
     case platformVersion
+
+    /**
+     * The version of the Stitch SDK that the device is currently using.
+     */
     case sdkVersion
 }
