+++ StitchCore/Sources/StitchCore/Auth/Internal/Models/APIAuthInfoImpl.swift
@@ -1,5 +1,8 @@
 import Foundation
 
+/**
+ * A struct containing the fields returned by the Stitch client API in an authentication request.
+ */
 public struct APIAuthInfoImpl: APIAuthInfo {
     enum CodingKeys: String, CodingKey {
         case userId = "user_id"
@@ -8,19 +11,43 @@ public struct APIAuthInfoImpl: APIAuthInfo {
         case refreshToken = "refresh_token"
     }
 
+    /**
+     * The ID of the newly authenticated user.
+     */
     public let userId: String
 
+    /**
+     * The device ID of the newly authenticated user.
+     */
     public let deviceId: String
 
+    /**
+     * The temporary access token for the newly authenticated user.
+     */
     public let accessToken: String
 
+    /**
+     * The permanent access token for the newly authenticated user.
+     */
     public let refreshToken: String
 }
 
+/**
+ * A struct containing the extended authentication information required by the `ExtendedAuthInfo` protocol.
+ */
 public struct ExtendedAuthInfoImpl: ExtendedAuthInfo {
+    /**
+     * A string describing the type of authentication provider used to log in as this user.
+     */
     public let loggedInProviderType: String
 
+    /**
+     * A string describing the name of the authentication provider used to log in as this user.
+     */
     public let loggedInProviderName: String
 
+    /**
+     * A `StitchUserProfile` object describing this user.
+     */
     public let userProfile: StitchUserProfile
 }
