+++ StitchCore/Sources/StitchCore/Auth/StitchCredential.swift
@@ -1,21 +1,18 @@
 import ExtendedJSON
 
 /**
- * A credential which can be used to log in as a Stitch user.
- * There is an implementation for each authentication provider
- * available in MongoDB Stitch. These implementations can be
- * generated using an authentication provider client.
+ * A credential which can be used to log in as a Stitch user. There is an implementation for each authentication
+ * provider available in MongoDB Stitch. These implementations can be generated using an authentication provider
+ * client.
  */
 public protocol StitchCredential {
     /**
-     * The name of the authentication provider for that
-     * this credential will be used to authenticate with.
+     * The name of the authentication provider that this credential will be used to authenticate with.
      */
     var providerName: String { get }
 
     /**
-     * The type of the authentication provider for that
-     * this credential will be used to authenticate with.
+     * The type of the authentication provider that this credential will be used to authenticate with.
      */
     var providerType: String { get }
 
@@ -23,8 +20,7 @@ public protocol StitchCredential {
     var material: Document { get }
 
     /**
-     * A `ProviderCapabilities` struct describing the behavior
-     * of this credential when logging in.
+     * A `ProviderCapabilities` struct describing the behavior of this credential when logging in.
      */
     var providerCapabilities: ProviderCapabilities { get }
 }
