+++ StitchCore/Sources/StitchCore/Internal/Net/Headers.swift
@@ -4,6 +4,9 @@ public enum Headers: String {
     case authorization = "Authorization"
     case bearer = "Bearer"
 
+    /**
+     * Returns an Authorization bearer header with the `forValue` parameter as the Bearer value.
+     */
     public static func authorizationBearer(forValue value: String) -> String {
         return "\(Headers.bearer.rawValue) \(value)"
     }
