+++ StitchCore/Sources/StitchCore/Internal/Net/Method.swift
@@ -1,3 +1,6 @@
+/**
+ * An enumeration of the available HTTP methods.
+ */
 public enum Method: String {
     case get = "GET",
     post = "POST",
