+++ StitchCore/Sources/StitchCore/StitchAppClientConfiguration.swift
@@ -1,20 +1,69 @@
+/**
+ * Properties representing the configuration of a client that communicate with a particular MongoDB Stitch application.
+ */
 public protocol StitchAppClientConfiguration: StitchClientConfiguration {
+    /**
+     * The client app ID of the Stitch application that this client is going to communicate with.
+     */
     var clientAppId: String { get }
+    
+    /**
+     * The name of the local application.
+     */
     var localAppName: String { get }
+    
+    /**
+     * The current version of the local application.
+     */
     var localAppVersion: String { get }
 }
 
-/** :nodoc: */
+/**
+ * :nodoc:
+ * The implementation of `StitchAppClientConfiguration`.
+ */
 public struct StitchAppClientConfigurationImpl: StitchAppClientConfiguration, Buildee {
+    /**
+     * The builder type that can build this configuration.
+     */
     public typealias TBuilder = StitchAppClientConfigurationBuilder
 
+    /**
+     * The base URL of the Stitch server that the client will communicate with.
+     */
     public let baseURL: String
+    
+    /**
+     * The underlying storage for authentication info.
+     */
     public let storage: Storage
+    
+    /**
+     * The `Transport` that the client will use to make round trips to the Stitch server.
+     */
     public let transport: Transport
+    
+    /**
+     * The client app ID of the Stitch application that this client is going to communicate with.
+     */
     public let clientAppId: String
+    
+    /**
+     * The name of the local application.
+     */
     public let localAppName: String
+    
+    /**
+     * The current version of the local application.
+     */
     public let localAppVersion: String
 
+    /**
+     * Initializes this configuration by accepting a `StitchAppClientConfigurationBuilder`.
+     *
+     * - throws: `StitchClientConfigurationError` or `StitchAppClientConfigurationError` if the builder is missing any
+     *           properties.
+     */
     public init(_ builder: TBuilder) throws {
         guard let clientAppId = builder.clientAppId else {
             throw StitchAppClientConfigurationError.missingClientAppId
@@ -52,28 +101,63 @@ public struct StitchAppClientConfigurationImpl: StitchAppClientConfiguration, Bu
     }
 }
 
-/** :nodoc: */
+/**
+ * An error that a Stitch app client configuration can throw if it is missing certain properties.
+ */
 public enum StitchAppClientConfigurationError: Error {
     case missingClientAppId
 }
 
+/**
+ * A builder that can build a `StitchAppClientConfiguration` object.
+ */
 public struct StitchAppClientConfigurationBuilder: StitchClientConfigurationBuilder, Builder {
+    /**
+     * :nodoc:
+     * configuration type that this builder builds.
+     */
     public typealias TBuildee = StitchAppClientConfigurationImpl
 
+    /**
+     * The base URL of the Stitch server that the client will communicate with.
+     */
     public var baseURL: String?
 
+    /**
+     * The underlying storage for authentication info.
+     */
     public var storage: Storage?
 
+    /**
+     * The `Transport` that the client will use to make round trips to the Stitch server.
+     */
     public var transport: Transport?
 
+    /**
+     * The client app ID of the Stitch application that this client is going to communicate with.
+     */
     public var clientAppId: String?
+    
+    /**
+     * The name of the local application.
+     */
     public var localAppName: String?
+    
+    /**
+     * The current version of the local application.
+     */
     public var localAppVersion: String?
 
+    /**
+     * Initializes the builder with a closure that sets the builder's desired properties.
+     */
     public init(_ builder: (inout StitchAppClientConfigurationBuilder) -> Void) {
         builder(&self)
     }
 
+    /**
+     * Builds the `StitchAppClientConfiguration`.
+     */
     public func build() throws -> StitchAppClientConfigurationImpl {
         return try StitchAppClientConfigurationImpl.init(self)
     }
